(refactor) Fix clang compilation

+++ engine/src/gfx/cockpit.cpp
@@ -487,8 +487,12 @@ inline void DrawITTSLine( QVector fromLoc, QVector aimLoc, GFXColor linecolor=GF
     GFXEnable( SMOOTH );
     GFXBlendMode( SRCALPHA, INVSRCALPHA );
     const float verts[2 * 3] = {
-        fromLoc.x, fromLoc.y, fromLoc.z,
-        aimLoc.x,   aimLoc.y,  aimLoc.z,
+        static_cast<float>(fromLoc.x),
+        static_cast<float>(fromLoc.y),
+        static_cast<float>(fromLoc.z),
+        static_cast<float>(aimLoc.x),
+        static_cast<float>(aimLoc.y),
+        static_cast<float>(aimLoc.z),
     };
     GFXDraw( GFXLINE, verts, 2 );
     GFXDisable( SMOOTH );
@@ -550,15 +554,28 @@ void GameCockpit::DrawTargetBox(const Radar::Sensor& sensor)
         if (draw_line_to_targets_target && targets_target != NULL) {
             QVector ttLoc = targets_target->Position();
             const float verts[3 * 3] = {
-                myLoc.x, myLoc.y, myLoc.z,
-                Loc.x,   Loc.y,   Loc.z,
-                ttLoc.x, ttLoc.x, ttLoc.x,
+                static_cast<float>(myLoc.x),
+                static_cast<float>(myLoc.y),
+                static_cast<float>(myLoc.z),
+
+                static_cast<float>(Loc.x),
+                static_cast<float>(Loc.y),
+                static_cast<float>(Loc.z),
+
+                static_cast<float>(ttLoc.x),
+                static_cast<float>(ttLoc.x),
+                static_cast<float>(ttLoc.x),
             };
             GFXDraw( GFXLINESTRIP, verts, 3 );
         } else {
             const float verts[2 * 3] = {
-                myLoc.x, myLoc.y, myLoc.z,
-                Loc.x,   Loc.y,   Loc.z,
+                static_cast<float>(myLoc.x),
+                static_cast<float>(myLoc.y),
+                static_cast<float>(myLoc.z),
+
+                static_cast<float>(Loc.x),
+                static_cast<float>(Loc.y),
+                static_cast<float>(Loc.z),
             };
             GFXDraw( GFXLINESTRIP, verts, 2 );
         }
